requirements.txt
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click
dnspython==2.7.0
docker==4.3.1
dockerpty==0.4.1
docopt==0.6.2
email_validator==2.2.0
Faker==33.3.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.16.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gitdb==4.0.12
GitPython==3.1.7
google-auth==2.37.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kubernetes==11.0.0
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
popper
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pykwalify==1.7.0
python-box==5.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==5.3.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
semver==3.0.2
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
soupsieve==2.6
spython==0.0.79
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wtf==0.1
WTForms==3.2.1
